agent/requirements.txt
flask==3.0.0
requests==2.31.0
gunicorn==22.0.0
pytest==8.2.0
python-dotenv==0.21.0